requirements.txt
# LLM
anthropic>=0.40.0

# Extraction PDF
pdfplumber>=0.11.0
python-docx>=1.1.0

# Base de données
sqlalchemy>=2.0.0

# Utilitaires
python-dotenv>=1.0.0
pydantic>=2.0.0

# Dashboard
streamlit>=1.40.0

# Tests
pytest>=8.0.0
